Replace individual shifts with direction-generic functions.
The wiki listed them as separate functions, but the optimizations
performed by the compiler make a single function more desirable. It's
now just a single copy of the code, and the compiler generates code for
constant direction parameters that is basically equivalent to the code
generated for the respective individual functions.